+++ R/meta-aux.R
@@ -1,9 +1,10 @@
-## Auxiliary functions
-##
-## Package: meta
-## Author: Guido Schwarzer <guido.schwarzer@uniklinik-freiburg.de>
-## License: GPL (>= 2)
-##
+# Auxiliary functions
+#
+# Package: meta
+# Author: Guido Schwarzer <guido.schwarzer@uniklinik-freiburg.de>
+# License: GPL (>= 2)
+#
+
 bylevs <- function(x) {
   if (is.factor(x))
     res <- levels(factor(x))
@@ -11,53 +12,59 @@ bylevs <- function(x) {
     res <- unique(x)
   res
 }
+
 byvarname <- function(argname, matchcall) {
-  ##
-  ## Determine name of subgroup variable
-  ##
+  #
+  # Determine name of subgroup variable
+  #
   res <- as.character(matchcall[[match(argname, names(matchcall))]])
-  ##
+  #
   if (length(res) > 1 & res[1] == "$")
     res <- res[length(res)]
-  ##
+  #
   if (length(res) == 0 || length(res) > 1)
     res <- "subgroup"
-  ##
+  #
   res
 }
+
 catch <- function(argname, matchcall, data, encl) {
-  ##
-  ## Catch value for argument
-  ##
+  #
+  # Catch value for argument
+  #
   eval(matchcall[[match(argname, names(matchcall))]], data, enclos = encl)
 }
+
 int2num <- function(x) {
-  ##
-  ## Convert integer to numeric
-  ##
+  #
+  # Convert integer to numeric
+  #
   if (is.integer(x))
     res <- as.numeric(x)
   else
     res <- x
-  ##
+  #
   res
 }
+
 npn <- function(x) {
-  ##
-  ## Check for non-positive values in vector
-  ##
+  #
+  # Check for non-positive values in vector
+  #
   selNA <- is.na(x)
   res <- selNA
   if (sum(!selNA) > 0)
     x[!selNA] <- x[!selNA] <= 0
-  ##
+  #
   res
 }
+
 replaceNULL <- function(x, replace = NA) {
   if (is.null(x))
     return(replace)
   x
 }
+
 replaceNA <- function(x, replace = NA) {
   if (is.null(x))
     return(x)
@@ -65,6 +72,7 @@ replaceNA <- function(x, replace = NA) {
     x[is.na(x)] <- replace
   x
 }
+
 warnarg <- function(x, y, fun, cl, otherarg) {
   if (x %in% y)
     if (!missing(cl))
@@ -76,9 +84,10 @@ warnarg <- function(x, y, fun, cl, otherarg) {
               "' in R function ", fun, ".\nSee help page of R function ",
               fun, " for information on the use of the new argument.",
               call. = FALSE)
-  ##
+  #
   invisible(NULL)
 }
+
 catchvar <- function(varname, x, mf) {
   res <- NULL
   error <-
@@ -86,14 +95,15 @@ catchvar <- function(varname, x, mf) {
                     x,
                     enclos = sys.frame(sys.parent())),
         silent = TRUE)
-  ##
+  #
   if (inherits(error, "try-error")) {
     res <- eval(mf[[match(varname, names(mf))]],
                 x$data, enclos = NULL)
   }
-  ##
+  #
   res
 }
+
 augment <- function(x, len, fun) {
   if (length(x) > 1)
     chklength(x, len, fun)
@@ -101,23 +111,25 @@ augment <- function(x, len, fun) {
     x <- rep(x, len)
   x
 }
+
 stoponly <- function(arg, val, func)
   stop("Argument ", arg, " =\"", val, "\"",
        " only defined for meta-analysis conducted with ",
        func, ".",
        call. = FALSE)
+
 deprecated <- function(newvar, newmiss, args, old, warn = TRUE) {
-  ##
+  #
   new <- deparse(substitute(newvar))
-  ##
+  #
   if (length(args) == 0)
     return(newvar)
-  ##
+  #
   if (is.list(args[[1]]))
     args <- args[[1]]
-  ##
+  #
   additional.arguments <- names(args)
-  ##
+  #
   if (!is.na(charmatch(old, additional.arguments)))
     if (!newmiss) {
       if (warn)
@@ -136,11 +148,12 @@ deprecated <- function(newvar, newmiss, args, old, warn = TRUE) {
   else
     return(newvar)
 }
+
 deprecated2 <- function(newvar, newmiss, oldvar, oldmiss, warn = TRUE) {
-  ##
+  #
   new <- deparse(substitute(newvar))
   old <- deparse(substitute(oldvar))
-  ##
+  #
   if (newmiss & oldmiss)
     return(newvar)
   else if (!newmiss & oldmiss)
@@ -160,6 +173,7 @@ deprecated2 <- function(newvar, newmiss, oldvar, oldmiss, warn = TRUE) {
     return(oldvar)
   }
 }
+
 runNN <- function(func, args, warn = TRUE) {
   args <- args[!sapply(args, is.null)]
   if (warn)
@@ -206,7 +220,7 @@ phi <- function(x) {
 # - 'vcov.basic' = variance-covariance matrix of the estimates.
 #
 # Returns all NMA estimates (basic + functional) and their standard errors.
-
+#
 basic2all <- function(trts, TE.basic, vcov.basic) {
   
   n.trts <- length(trts)
@@ -269,4 +283,138 @@ basic2all <- function(trts, TE.basic, vcov.basic) {
   res <- list(TEs = TEs, seTEs = seTEs)
   #
   res
+}
+
+setNA <- function(x) {
+  x[!is.na(x)] <- NA
+  x  
+}
+
+setNA_subset <- function(x, subset) {
+  x[subset] <- NA
+  x
+}
+
+setNA_vars <- function(x, varnames) {
+  x[names(x) %in% varnames] <- NA
+  x
+}
+
+setNA_nma <- function(x) {
+   x$seTE.adj <- setNA(x$seTE.adj)
+   x$seTE.adj.common <- setNA(x$seTE.adj.common)
+   x$seTE.adj.random <- setNA(x$seTE.adj.random)
+   x$correlated <- setNA(x$correlated)
+   #
+   x$TE.nma.common <- setNA(x$TE.nma.common)
+   x$seTE.nma.common <- setNA(x$seTE.nma.common)
+   x$lower.nma.common <- setNA(x$lower.nma.common)
+   x$upper.nma.common <- setNA(x$upper.nma.common)
+   x$statistic.nma.common <- setNA(x$statistic.nma.common)
+   x$pval.nma.common <- setNA(x$pval.nma.common)
+   #
+   x$leverage.common <- setNA(x$leverage.common)
+   x$w.common <- setNA(x$w.common)
+   x$Q.common <- setNA(x$Q.common)
+   #
+   x$statistic.common <- setNA(x$statistic.common)
+   x$pval.common <- setNA(x$pval.common)
+   #
+   x$TE.nma.random <- setNA(x$TE.nma.random)
+   x$seTE.nma.random <- setNA(x$seTE.nma.random)
+   x$lower.nma.random <- setNA(x$lower.nma.random)
+   x$upper.nma.random <- setNA(x$upper.nma.random)
+   x$statistic.nma.random <- setNA(x$statistic.nma.random)
+   x$pval.nma.random <- setNA(x$pval.nma.random)
+   #
+   x$w.random <- setNA(x$w.random)
+   #
+   x$statistic.random <- setNA(x$statistic.random)
+   x$pval.random <- setNA(x$pval.random)
+   #
+   x$seTE.predict <- setNA(x$seTE.predict)
+   x$lower.predict <- setNA(x$lower.predict)
+   x$upper.predict <- setNA(x$upper.predict)
+   #
+   x$prop.direct.common <- setNA(x$prop.direct.common)
+   x$prop.direct.random <- setNA(x$prop.direct.random)
+   #
+   x$TE.indirect.common <- setNA(x$TE.indirect.common)
+   x$seTE.indirect.common <- setNA(x$seTE.indirect.common)
+   x$lower.indirect.common <- setNA(x$lower.indirect.common)
+   x$upper.indirect.common <- setNA(x$upper.indirect.common)
+   x$statistic.indirect.common <- setNA(x$statistic.indirect.common)
+   x$pval.indirect.common <- setNA(x$pval.indirect.common)
+   #
+   x$TE.indirect.random <- setNA(x$TE.indirect.random)
+   x$seTE.indirect.random <- setNA(x$seTE.indirect.random)
+   x$lower.indirect.random <- setNA(x$lower.indirect.random)
+   x$upper.indirect.random <- setNA(x$upper.indirect.random)
+   x$statistic.indirect.random <- setNA(x$statistic.indirect.random)
+   x$pval.indirect.random <- setNA(x$pval.indirect.random)
+   #
+   x$Q <- setNA(x$Q)
+   x$df.Q <- setNA(x$df.Q)
+   x$pval.Q <- setNA(x$pval.Q)
+   x$I2 <- setNA(x$I2)
+   x$lower.I2 <- setNA(x$lower.I2)
+   x$upper.I2 <- setNA(x$upper.I2)
+   x$tau <- setNA(x$tau)
+   x$tau2 <- setNA(x$tau2)
+   #
+   x$Q.heterogeneity <- setNA(x$Q.heterogeneity)
+   x$df.Q.heterogeneity <- setNA(x$df.Q.heterogeneity)
+   x$pval.Q.heterogeneity <- setNA(x$pval.Q.heterogeneity)
+   #
+   x$Q.inconsistency <- setNA(x$Q.inconsistency)
+   x$df.Q.inconsistency <- setNA(x$df.Q.inconsistency)
+   x$pval.Q.inconsistency <- setNA(x$pval.Q.inconsistency)
+   #
+   x$Q.decomp <- setNA(x$Q.decomp)
+   #
+   x$W.matrix.common <- setNA(x$W.matrix.common)
+   x$W.matrix.random <- setNA(x$W.matrix.random)
+   #
+   x$L.matrix.common <- setNA(x$L.matrix.common)
+   x$Lplus.matrix.common <- setNA(x$Lplus.matrix.common)
+   x$L.matrix.random <- setNA(x$L.matrix.random)
+   x$Lplus.matrix.random <- setNA(x$Lplus.matrix.random)
+   #
+   x$Q.matrix <- setNA(x$Q.matrix)
+   #
+   x$G.matrix <- setNA(x$G.matrix)
+   #
+   x$H.matrix.common <- setNA(x$H.matrix.common)
+   x$H.matrix.random <- setNA(x$H.matrix.random)
+   #
+   x$P.common <- setNA(x$P.common)
+   x$P.random <- setNA(x$P.random)
+   #
+   x$Cov.common <- setNA(x$Cov.common)
+   x$Cov.random <- setNA(x$Cov.random)
+   #
+   x
+}
+
+
+
+if (FALSE) {
+  TE.direct.common = TE.direct
+  seTE.direct.common = seTE.direct
+  lower.direct.common = lower.direct
+  upper.direct.common = upper.direct
+  statistic.direct.common = statistic.direct
+  pval.direct.common = pval.direct
+  #
+  TE.direct.random = res.r$TE.direct
+  seTE.direct.random = res.r$seTE.direct
+  lower.direct.random = res.r$lower.direct
+  upper.direct.random = res.r$upper.direct
+  statistic.direct.random = res.r$statistic.direct
+  pval.direct.random = res.r$pval.direct
+  #
+  Q.direct = res.r$Q.direct
+  tau.direct = sqrt(res.r$tau2.direct)
+  tau2.direct = res.r$tau2.direct
+  I2.direct = res.r$I2.direct
 }