+++ R/meta-chk.R
@@ -253,11 +253,16 @@ chkdeprecated <- function(x, new, old, warn = TRUE) {
 }
 
 chksuitable <- function(x, method,
-                        classes = "netmetabin",
+                        classes = c("netmetabin",
+                                    "netmeta.crossnma",
+                                    "netmeta.multinma"),
                         addtext = "",
                         stop = TRUE,
                         status = "implemented") {
   func <- if (stop) base::stop else base::warning
+  #
+  if (length(addtext) != length(classes) & length(addtext) == 1)
+    addtext <- rep_len(addtext, length(classes))
   #
   for (i in seq_along(classes)) {
     if (inherits(x, classes[i])) {