+++ R/netcomb.R
@@ -381,6 +381,8 @@ netcomb <- function(x,
   ##
   ##
   chkclass(x, "netmeta")
+  chksuitable(x, "Component network meta-analysis",
+              classes = c("netmeta.crossnma", "netmeta.multinma"))
   ##
   x <- updateversion(x)
   ##
@@ -567,11 +569,11 @@ netcomb <- function(x,
                                 grepl,
                                 pattern = unident.pattern), any))]
     ##
-    res.c$components <- lapply(res.c$components, setNA, comps.unident)
-    res.c$combinations <- lapply(res.c$combinations, setNA, unident.combs)
+    res.c$components <- lapply(res.c$components, setNA_vars, comps.unident)
+    res.c$combinations <- lapply(res.c$combinations, setNA_vars, unident.combs)
     ##
-    res.r$components <- lapply(res.r$components, setNA, comps.unident)
-    res.r$combinations <- lapply(res.r$combinations, setNA, unident.combs)
+    res.r$components <- lapply(res.r$components, setNA_vars, comps.unident)
+    res.r$combinations <- lapply(res.r$combinations, setNA_vars, unident.combs)
   }
   ##
   if (length(comps.unident) == 0)